# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AesthTech Solutions API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  role: UserProfileRole;
  isPremium: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileBody {
  name?: string;
  bio?: string;
  phone?: string;
  avatarUrl?: string;
}

export interface UserStats {
  assessmentsCompleted: number;
  challengesAttempted: number;
  totalScore: number;
  /** @nullable */
  rank?: number | null;
  streakDays: number;
  isPremium: boolean;
}

export interface Assessment {
  id: number;
  title: string;
  description: string;
  category: string;
  questionCount: number;
  estimatedMinutes: number;
  isPremium: boolean;
}

export interface StartAssessmentBody {
  assessmentId: number;
}

export interface AssessmentQuestion {
  id: number;
  text: string;
  options: string[];
  order: number;
}

export type AssessmentSessionStatus =
  (typeof AssessmentSessionStatus)[keyof typeof AssessmentSessionStatus];

export const AssessmentSessionStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface AssessmentSession {
  id: number;
  assessmentId: number;
  assessmentTitle: string;
  status: AssessmentSessionStatus;
  questions: AssessmentQuestion[];
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type SubmitAssessmentBodyAnswersItem = {
  questionId: number;
  answer: number;
};

export interface SubmitAssessmentBody {
  answers: SubmitAssessmentBodyAnswersItem[];
}

export type AssessmentResultCareerMatchesItem = {
  career: string;
  matchPercent: number;
};

export interface AssessmentResult {
  sessionId: number;
  score: number;
  totalQuestions: number;
  recommendations: string[];
  careerMatches: AssessmentResultCareerMatchesItem[];
  completedAt: string;
}

export type ChallengeDifficulty =
  (typeof ChallengeDifficulty)[keyof typeof ChallengeDifficulty];

export const ChallengeDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Challenge {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: ChallengeDifficulty;
  timeLimit: number;
  questionCount: number;
  isPremium: boolean;
  attempts: number;
}

export interface ChallengeQuestion {
  id: number;
  text: string;
  options: string[];
  order: number;
}

export type ChallengeWithQuestionsDifficulty =
  (typeof ChallengeWithQuestionsDifficulty)[keyof typeof ChallengeWithQuestionsDifficulty];

export const ChallengeWithQuestionsDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface ChallengeWithQuestions {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: ChallengeWithQuestionsDifficulty;
  timeLimit: number;
  isPremium: boolean;
  questions: ChallengeQuestion[];
}

export type ChallengeAttemptBodyAnswersItem = {
  questionId: number;
  answer: number;
};

export interface ChallengeAttemptBody {
  answers: ChallengeAttemptBodyAnswersItem[];
  timeTaken: number;
}

export interface ChallengeAttemptResult {
  id: number;
  challengeId: number;
  challengeTitle: string;
  score: number;
  totalQuestions: number;
  timeTaken: number;
  /** @nullable */
  rank?: number | null;
  isPassed: boolean;
  completedAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  userName: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalScore: number;
  challengesCompleted: number;
}

export type CreateOrderBodyPlanId =
  (typeof CreateOrderBodyPlanId)[keyof typeof CreateOrderBodyPlanId];

export const CreateOrderBodyPlanId = {
  monthly: "monthly",
  annual: "annual",
  lifetime: "lifetime",
} as const;

export interface CreateOrderBody {
  planId: CreateOrderBodyPlanId;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  razorpayKeyId: string;
  planId: string;
}

export interface VerifyPaymentBody {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
  planId: string;
}

export interface PaymentVerification {
  success: boolean;
  message: string;
  isPremium: boolean;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface Payment {
  id: number;
  userId: number;
  razorpayOrderId: string;
  /** @nullable */
  razorpayPaymentId?: string | null;
  amount: number;
  currency: string;
  status: PaymentStatus;
  planId: string;
  createdAt: string;
}

export interface Subscription {
  /** @nullable */
  id?: number | null;
  /** @nullable */
  planId?: string | null;
  isPremium: boolean;
  /** @nullable */
  expiresAt?: string | null;
  status: string;
}

export interface ContactBody {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface NewsletterBody {
  email: string;
}

export interface AdminStats {
  totalUsers: number;
  premiumUsers: number;
  totalRevenue: number;
  assessmentsCompleted: number;
  challengesAttempted: number;
  recentSignups: number;
  activeSubscriptions: number;
}

export interface AdminUsersResponse {
  users: UserProfile[];
  total: number;
  page: number;
}

export type ListChallengesParams = {
  difficulty?: ListChallengesDifficulty;
  category?: string;
};

export type ListChallengesDifficulty =
  (typeof ListChallengesDifficulty)[keyof typeof ListChallengesDifficulty];

export const ListChallengesDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type ListAdminUsersParams = {
  page?: number;
  limit?: number;
};