# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AesthTech Solutions API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List all career assessments
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"

    post:
      operationId: startAssessment
      tags: [assessments]
      summary: Start a new assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAssessmentBody"
      responses:
        "201":
          description: Assessment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentSession"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assessments/{id}/submit:
    post:
      operationId: submitAssessment
      tags: [assessments]
      summary: Submit assessment answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssessmentBody"
      responses:
        "200":
          description: Assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List interview challenges
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get challenge by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeWithQuestions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /challenges/{id}/attempt:
    post:
      operationId: submitChallengeAttempt
      tags: [challenges]
      summary: Submit challenge attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeAttemptBody"
      responses:
        "200":
          description: Attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeAttemptResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        isPremium:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isPremium
        - createdAt
...
    Assessment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        questionCount:
          type: integer
        estimatedMinutes:
          type: integer
        isPremium:
          type: boolean
      required:
        - id
        - title
        - description
        - category
        - questionCount
        - estimatedMinutes
        - isPremium
    StartAssessmentBody:
      type: object
      properties:
        assessmentId:
          type: integer
      required:
        - assessmentId
    AssessmentQuestion:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        order:
          type: integer
      required:
        - id
        - text
        - options
        - order
    AssessmentSession:
      type: object
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        assessmentTitle:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentQuestion"
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - assessmentId
        - assessmentTitle
        - status
        - questions
        - startedAt
    SubmitAssessmentBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              answer:
                type: integer
            required:
              - questionId
              - answer
      required:
        - answers
    AssessmentResult:
      type: object
      properties:
        sessionId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        recommendations:
          type: array
          items:
            type: string
        careerMatches:
          type: array
          items:
            type: object
            properties:
              career:
                type: string
              matchPercent:
                type: integer
            required:
              - career
              - matchPercent
        completedAt:
          type: string
      required:
        - sessionId
        - score
        - totalQuestions
        - recommendations
        - careerMatches
        - completedAt

    # Challenges
    Challenge:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        timeLimit:
          type: integer
        questionCount:
          type: integer
        isPremium:
          type: boolean
        attempts:
          type: integer
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - timeLimit
        - questionCount
        - isPremium
        - attempts
    ChallengeQuestion:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        order:
          type: integer
      required:
        - id
        - text
        - options
        - order
    ChallengeWithQuestions:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        timeLimit:
          type: integer
        isPremium:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeQuestion"
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - timeLimit
        - isPremium
        - questions
    ChallengeAttemptBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              answer:
                type: integer
            required:
              - questionId
              - answer
        timeTaken:
          type: integer
      required:
        - answers
        - timeTaken
    ChallengeAttemptResult:
      type: object
      properties:
        id:
          type: integer
        challengeId:
          type: integer
        challengeTitle:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        timeTaken:
          type: integer
        rank:
          type: ["integer", "null"]
        isPassed:
          type: boolean
        completedAt:
          type: string
      required:
        - id
        - challengeId
        - challengeTitle
        - score
        - totalQuestions
        - timeTaken
        - isPassed
        - completedAt
...