# lib/db/src/schema/assessments.ts
import { pgTable, text, serial, timestamp, boolean, integer, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const assessmentsTable = pgTable("assessments", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  description: text("description").notNull(),
  category: text("category").notNull(),
  questionCount: integer("question_count").notNull().default(10),
  estimatedMinutes: integer("estimated_minutes").notNull().default(15),
  isPremium: boolean("is_premium").notNull().default(false),
  questions: jsonb("questions").notNull().default([]),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const assessmentSessionsTable = pgTable("assessment_sessions", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull(),
  assessmentId: integer("assessment_id").notNull(),
  status: text("status").notNull().default("in_progress"),
  answers: jsonb("answers"),
  score: integer("score"),
  recommendations: jsonb("recommendations"),
  careerMatches: jsonb("career_matches"),
  startedAt: timestamp("started_at", { withTimezone: true }).notNull().defaultNow(),
  completedAt: timestamp("completed_at", { withTimezone: true }),
});

export const insertAssessmentSchema = createInsertSchema(assessmentsTable).omit({ id: true, createdAt: true });
export type InsertAssessment = z.infer<typeof insertAssessmentSchema>;
export type Assessment = typeof assessmentsTable.$inferSelect;

export const insertAssessmentSessionSchema = createInsertSchema(assessmentSessionsTable).omit({ id: true, startedAt: true });
export type InsertAssessmentSession = z.infer<typeof insertAssessmentSessionSchema>;
export type AssessmentSession = typeof assessmentSessionsTable.$inferSelect;